+++ packages/apollo-server-core/src/plugin/landingPage/default/index.ts
@@ -15,12 +15,25 @@ export interface ApolloServerPluginLandingPageDefaultBaseOptions {
    */
   footer?: boolean;
   /**
-   * Folks can configure their landing page to link to Studio Explorer with a
-   * document, variables and headers loaded in the UI.
+   * Users can configure their landing page to link to Studio Explorer with a
+   * document loaded in the UI.
    */
   document?: string;
+  /**
+   * Users can configure their landing page to link to Studio Explorer with
+   * variables loaded in the UI.
+   */
   variables?: Record<string, string>;
+  /**
+   * Users can configure their landing page to link to Studio Explorer with
+   * headers loaded in the UI.
+   */
   headers?: Record<string, string>;
+  /**
+   * Users can configure their landing page to link to Studio Explorer with the
+   * setting to include/exclude cookies loaded in the UI.
+   */
+  includeCookies?: boolean;
   // For Apollo use only.
   __internal_apolloStudioEnv__?: 'staging' | 'prod';
 }
@@ -47,6 +60,7 @@ interface LandingPageConfig {
   document?: string;
   variables?: Record<string, string>;
   headers?: Record<string, string>;
+  includeCookies?: boolean;
   footer?: boolean;
 }
 
@@ -56,13 +70,26 @@ export function ApolloServerPluginLandingPageLocalDefault(
   // We list known keys explicitly to get better typechecking, but we pass
   // through extras in case we've added new keys to the splash page and haven't
   // quite updated the plugin yet.
-  const { version, __internal_apolloStudioEnv__, footer, ...rest } = options;
+  const {
+    version,
+    __internal_apolloStudioEnv__,
+    footer,
+    document,
+    variables,
+    headers,
+    includeCookies,
+    ...rest
+  } = options;
   return ApolloServerPluginLandingPageDefault(
     version,
     encodeConfig({
       isProd: false,
       apolloStudioEnv: __internal_apolloStudioEnv__,
       footer,
+      document,
+      variables,
+      headers,
+      includeCookies,
       ...rest,
     }),
   );
@@ -74,14 +101,27 @@ export function ApolloServerPluginLandingPageProductionDefault(
   // We list known keys explicitly to get better typechecking, but we pass
   // through extras in case we've added new keys to the splash page and haven't
   // quite updated the plugin yet.
-  const { version, __internal_apolloStudioEnv__, footer, graphRef, ...rest } =
-    options;
+  const {
+    version,
+    __internal_apolloStudioEnv__,
+    footer,
+    document,
+    variables,
+    headers,
+    includeCookies,
+    graphRef,
+    ...rest
+  } = options;
   return ApolloServerPluginLandingPageDefault(
     version,
     encodeConfig({
       isProd: true,
       apolloStudioEnv: __internal_apolloStudioEnv__,
       footer,
+      document,
+      variables,
+      headers,
+      includeCookies,
       graphRef,
       ...rest,
     }),